update text editor, remove formik prop on meta field component
