+++ src/components/inputs/formik-text-editor.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { TextEditor } from "openstack-uicore-foundation/lib/components";
+import TextEditorV3 from "openstack-uicore-foundation/lib/components/inputs/editor-input-v3";
 import { useFormikContext } from "formik";
 
 const FormikTextEditor = ({ name, ...props }) => {
@@ -7,13 +7,14 @@ const FormikTextEditor = ({ name, ...props }) => {
     useFormikContext();
 
   return (
-    <TextEditor
+    <TextEditorV3
       name={name}
       id={name}
       value={values[name]}
       onChange={(e) => setFieldValue(name, e.target.value)}
       onBlur={() => setFieldTouched(name, true)}
       error={touched?.[name] && errors?.[name] ? errors?.[name] : ""}
+      license={process.env.JODIT_LICENSE_KEY}
       {...props}
     />
   );
